issues : move github specific stuff to separate module
