+++ Makefile
@@ -35,6 +35,7 @@ ghcli_SRCS			=	src/ghcli.c			\
 					src/github/comments.c		\
 					src/github/forks.c		\
 					src/github/pulls.c		\
+					src/github/issues.c		\
 					thirdparty/sn/sn.c		\
 					thirdparty/pdjson/pdjson.c
